Evaluation of the validation process results
The results of the validation process now will be added to each element
as a dynamic array. It contains a list of all matching (valid) asn and
each asn in turn includes a list of all corresponding prefixes. The
values will be printed in a relevant form
(asn1-pfx1-pfx2,asn2-pfx3-pfx4).

Also undo the origin-asn-field in the elem struct.